*Skilte ut komponent for visning av avvikende åpningstider til egen fil *Bugfix for disabling av chat utenfor åpningstid
