+++ src/App.less
@@ -165,3 +165,13 @@ body {
 .btn-auth {
   min-height: 1.5rem;
 }
+
+.apningstider-avvik {
+  p {
+    margin: 0;
+  }
+}
+
+.apningstider-avvik-header {
+  font-weight: bold;
+}
