+++ src/pages/chat/Chat.less
@@ -45,16 +45,6 @@
     }
   }
 
-  &__avvik {
-    p {
-      margin: 0;
-    }
-  }
-
-  &__avvik-header {
-    font-weight: bold;
-  }
-
   &__panel-start-knapp {
     text-align: center;
     margin-top: 2em;
