+++ src/pages/chat/ChatTemasideBase.tsx
@@ -9,52 +9,16 @@ import { Hovedknapp } from "nav-frontend-knapper";
 import { urls, vars } from "../../Config";
 import PanelBase from "nav-frontend-paneler";
 import { AlertStripeInfo } from "nav-frontend-alertstriper";
-import Apningstider from "../../utils/apningstider";
 import { fetchServerTidOffset } from "../../clients/apiClient";
 import { logEvent } from "../../utils/logger";
-import FormattedMsgMedParagrafer from "../../components/intl-msg-med-paragrafer/FormattedMsgMedParagrafer";
-import moment from "moment-timezone";
+import ApningstiderAvvik from "../../components/apningstider/ApningstiderAvvik";
 
 type ChatTemaProps = {
   chatTemaData: ChatTemaData,
   children: ReactNode
 };
 
-type AvvikProps = {
-  apningstider: Apningstider,
-  harChatbot?: boolean
-};
-
 const cssPrefix = "chat-tema";
-
-const ApningstiderAvvik = ({apningstider, harChatbot = false}: AvvikProps) => {
-  const apningstiderAvvik = apningstider.getAktuelleAvvikstider();
-  const datoFormat = (dato: string) => moment(dato, "DD-MM-YYYY").format("DD.MM.");
-
-  return apningstiderAvvik.length > 0 ? (
-    <div className={`${cssPrefix}__avvik`}>
-      <span className={`${cssPrefix}__avvik-header`}>
-        <FormattedMsgMedParagrafer id={harChatbot ? "apningstid.avvik.chatbot" : "apningstid.avvik"} />
-      </span>
-      {apningstiderAvvik.map((datoTidsrom, index) => (
-        datoTidsrom.tidsrom ? (
-            <FormattedMsgMedParagrafer
-              id="apningstid.avvik.apent"
-              values={{dato: datoFormat(datoTidsrom.dato), start: datoTidsrom.tidsrom.start, end: datoTidsrom.tidsrom.end}}
-              key={`tid${index}`}
-            />
-          )
-          : (
-            <FormattedMsgMedParagrafer
-              id="apningstid.avvik.stengt"
-              values={{dato: datoFormat(datoTidsrom.dato)}}
-              key={`tid${index}`}
-            />
-          )
-      ))}
-    </div>
-  ) : null;
-};
 
 const ChatTemaSideBase = ({ chatTemaData, children }: ChatTemaProps) => {
   const [chatButtonClicked, setChatButtonClicked] = useState();
